New version of TSpectrum2::Search2General Now the algorithm is based on 2-nd, 4-th, 6-th smoothed differences.The meaning of the threshold is the same as in 1-dimensional case. The sigma (beacause of 6-th order differences) should not be less than 2. Recommended values are 4 and higher.
git-svn-id: http://root.cern.ch/svn/root/trunk@6577 27541ba8-7e3a-0410-8455-c3a389f83636

+++ hist/inc/TSpectrum2.h
@@ -1,4 +1,4 @@
-// @(#)root/hist:$Name:  $:$Id: TSpectrum.h,v 1.5 2001/06/22 16:10:17 rdm Exp $
+// @(#)root/hist:$Name:  $:$Id: TSpectrum2.h,v 1.1 2003/04/14 12:56:34 brun Exp $
 // Author: Miroslav Morhac   27/05/99
 
 /*************************************************************************
@@ -227,7 +227,7 @@ public:
    const char *Smooth2(float **spectrum,int sizex,int sizey,int pointsx,int pointsy);
    void DecFourier2(double *working_space,int num,int iter,int inv);
    const char *Deconvolution2HighResolution(float** source,const float** resp,int sizex,int sizey,int number_of_iterations,int number_of_repetitions,double boost);
-   Int_t Search2General(float **source,int sizex,int sizey,double sigma,int threshold,bool markov,int aver_window);
+   Int_t Search2General(float **source,int sizex,int sizey,double sigma,double threshold,bool markov,int aver_window);
    double      Lls(double a);
 
       //auxiliary functions for 2. parameter fit functions
